basic user registration flow including forwarding request to backend via soap

+++ agent/CarRentalPortal.Application/Users/Commands/CreateUser/CreateUserSoapCommand.cs
@@ -2,7 +2,7 @@
 
 namespace CarRentalPortal.Application.Users.Commands.CreateUser
 {
-    public class CreateUserSoapCommand : IRequest<int>
+    public class CreateUserSoapCommand : IRequest<long>
     {
         public string Username { get; set; }
         public string Email { get; set; }
